First pass at cleaning up gateway docs
As was pointed out by a certain unamed twitch streamer (albeit in a fairly rude and nonconstructive format), certain parts of the gateway documentation have rough edges. The gateway docs where the last part of our documentation I wrote, and ended up fairly rushed. At the time I did not have a great understand of various important aspects of our connection process, and the various additions/iterations to this documentation has not aided in clearing any of those up. This commit is the first in an effort to clean these docs up, and bring them more or less up to par with the rest of our docs, with a focus on making them more consumable for the average inexperienced user.