Add permissions calculator link
This gem appears only in the Discord API server faq, and I
thought it should also appear in the docs because it's
really useful.

I wasn't sure if the authors' names should appear here, but
I decided to omit it because their names appear on the
calculator when the window is large enough.