Add Response Code 50021
Can't seem to find any documentation or implementation of this.

```{"code": 50021, "message": "Cannot execute action on a system message"}```

For reference, this happened on an event to add a message reaction and hit a message that happened to be a pinned message notification ["user pinned a message to this channel. See all the pins."] - which is a system message that cannot have a reaction added to it.