Clarify presence condition of partial member field (#984)
Currently it is ambiguous when the partial member field in user objects
in message mentions fields are filled. This clarifies that it is only
filled in message events received from text-based guild channels. It
should not be, for example, expected to be filled in message objects
received from HTTP requests.

- The style of this clarification is based on the Guild Object structure
  table that uses asterisks to denote fields that are only filled in Guild
  Create events.

- The language "text-based gulid channels" is used to cover `GUILD_TEXT`,
  `GUILD_NEWS`, and any other future channel types that could be expected
  to yield message objects.