Add documentation for MessageFlags and MessageReference
`MessageFlags` indicate specific features of a message or how it should be handled, and is represented as an integer bit set in the `flags` property of a message object.

`MessageReference` is sent with crossposted messages indicating where the original message was sent, and can be used to link back to the originating message.