fix: remove ineffective mitigation from warning
Per #1241, messages with >250 special characters (in a row?) will have all special characters stripped from the message, including U-200B.  When this occurs, the U-200B placed between the "@" and "e" will also be stripped, making this method of mention mitigation ineffective.

While U-200B can be used to mitigate mentions if done with care, I think the complexity of this caveat means that most users will not implement effective mitigations, and thus be blindly vulnerable to the issue. Suggesting bot developers strip all "@" characters when an uncontrolled message is sent is a safer and more thorough solution.