Added buffer size note to gateway
The gateway provides no guarantee that it will actually chunk my messages up into multiple packets using transport compression. This may be problematic for bots that expect to be able to allocate a vector/array-based buffer. 

The first dozen packets from the gateway will likely be much larger than the rest of the packets we receive in the majority of cases.

Suppose we receive a 20MB message, we reallocate a buffer to support this, but having not expected this, we do not have code in place to reduce the buffer size afterwards. We now are consuming likely 10x more memory than generally needed to support this buffer. 

While this may be intuitive to library devs already familiar with the gateway, this is not helpful for new developers who may still be overwhelmed with the vast quantity of detail this documentation provides, so it is worth making a note of this just to provide helpful information.

I have also adjusted the syntax of the python example to be correct (for example, you cannot reassign a global name without declaring it as being global. Might as well make sure the examples provided are actually valid).