Updated 'Functions in the SDK' C# Binding Example  (#1579)
* Merge branch 'pr/15'

* Fixed a Lobby Example

Updated The SDK Way to reflect GameSDK changes.

* Removed unnessecary Github notions

* Updated 'Functions in the SDK' C# Binding Example

Changes the C# Binding example for returning normally to include the OnCurrentUserUpdate Function as GetCurrentUser will otherwise return null.
Prettier-ed the same OnCurrentUserUpdate Function in Users docs page. Re-submitting Lobby Example and Overlay Toggle Examples in this as well.

* Typo in Dispatch Credentials MacOs Example

Additional Typo has surfaced, this addresses that.

* Updated Example Networking,

Some typos with outdated code and a fetching a memberId is already an Id, not a user entirely.
Fixed example for networkManager.Flush as it wasn't in the proper function for LateUpdate!