Added alt text to images, moved GitHub CDN-hosted images into repo.
Fixed several bits of broken formatting, a bunch of NBSP characters
that are not used elsewhere, and fixed a few broken code examples
(missing semi-colons, et cetera).

Added a table describing the example of the Rich Presence activity
example, since it is not clear to users who have impaired sight as to
how this may look.

This is added in response to
https://github.com/discord/discord-api-docs/pull/2097#discussion_r495313517
on PR #2097 by Mason.