Document APIv9. (#2855)
To make migration easier, there is a guide in `docs/topics/Threads.md` that aggregates all the information about threads in one place for you!

We wanted to give app developers at least 2 months heads up so you all can update your bots and be ready to go whenever we do ship threads.  Please keep in mind that bots on older API versions will receive NO gateway events for threads, or messages in threads, so updating is important!

I'll be around for questions today, and I plan to keep this open for ~24 hours and then merge tomorrow afternoon and update the docs website.