Update version in the sample query-string, add some info about intents
This won't be of much importance but it would be good if the version in the sample query-string is v9, since it's already released. But, threads haven't been shipped yet so it's no use yet. Nonetheless, it's still a good practice to be up-to-date. 

Another thing about this commit is adding the info of the intent calculators, since the docs doesn't have any content telling users about them. Lastly, this improves the statement about intents being required.