Fixes erroneous tags description.
After a short discussion (#3562), it seems that the tags field can be any
string of text and is used in some way to power the autocomplete/suggestion popup.
The client using/displaying emoji names for the tag is just a good way of generating
a relevant keyword from a user, using the emojis as a pictorial aide, and is thus
not part of the API's concerns.