Tweak wording regarding command names and bulk overwrites.
Changes the wording surrounding command names to match observed behaviour
w.r.t upper case characters with case folding rules that produce other
upper case characters, and characters for which case makes no sense.

Changes the wording so that bulk overwrites cannot be confused for a bulk
itemwise overwrite, which they are not.