change the max and min ranges of the integer type
Application Command Option Type has an integer type, which the docs said was between -2^53 and 2^53. This is actually the correct range for `NUMBER`, a `double` type.
Testing and investigation reveals the correct range is between -2147483647 and 2147483647 making this type a signed 32 bit value.

I have corrected the docs to match the implementation in use.