Clarify initial Heartbeat delay
Currently, it has to be inferred that `random.random()` means [this Python API](https://docs.python.org/3/library/random.html#random.random). I'm open to input on how to improve this change, but having to infer language to go look up API docs to know the range of the multiplier is a lot for such a small detail.