Add discord-py-interactions to Community Resources
**discord-py-interactions** is a very simple API wrapper intended to exclusively make use of the new interactions released by Discord into their Public API.

Originally a "monkey-patch" and 3rd party addon to the well renowned public library *discord.py*, this library now aims to be fully independent of its parent library -- able to handle the basic necessities that a lib must require in order to function, e.g. the Gateway, sane HTTP requests, and dispatched event data as callbacks. These are some of the few things that this library is expected to do, but is also able to do much more than just that.

**Why use discord-interactions?**
Our library (not to be confused with the officially maintained version from Discord, `discord-interactions`) specializes in the implementation and integration of interactions in Python Discord bots. This is a strong suit that we wish to focus on and become the very best at because we strongly believe in interactions being the future of bots on Discord. We include, but are not limited to some core features such as:

- Dynamic object data generation via. dispatch
- Suggestive/implicit caching system
- Declarative and inferred callback decorators for components and commands
- Bridging support and compatibility between d.py forks and other Python libraries

**How do we handle HTTP rate limiting sanely?**
Please refer to this [section of code here](https://github.com/goverfl0w/discord-interactions/blob/unstable/interactions/api/http.py#L205-L217) where you can currently access our beta launch of the library.

**How does the client work?**
Please refer to this [section of code here](https://github.com/goverfl0w/discord-interactions/blob/unstable/interactions/client.py#L23).