remove json error code table
Remove the table of error codes mapped to their messages from the JSON
section of the "Opcodes and Status Codes" page. This table was
originally created when the message for an error code was not returned
in the response, but now that this is the case with API version 8 and up
we are removing the table of explicitly documented mappings.