Add note re user permissions and user interactions (#4626)
It is currently undocumented that a user must be able to "Send Messages"
as well as "Use Application Commands" in order to invoke a user command.
This comes as an additional complication because of the intersection of
channel permissions and user permissions.  Attempts to invoke a user
command in a channel which the user cannot speak in (e.g., a rules
channel) will result in a permission denied error, as if the underlying
application command provider had insufficient permissions.

This is compounded by no error (or, indeed, no request) ever being sent
to the underlying application command provider over the WS connection.