Clarify how user updates are dispatched
When a user is updated (eg. username or avatar), a GUILD_MEMBER_UPDATE is dispatched, as said by the docs.

However, the docs also indicate that a PRESENCE_UPDATE is triggered, which doesn't seem to be the case anymore.
I've updated the definition of PRESENCE_UPDATE to remove this information.

Additionally, I clarified that the USER_UPDATE event is dispatched only for updates to the current bot's user (this event isn't fired when other users are updated).