Updated "Message" and "Webhook" resource page and "Using Message Components" page with the new components v2 behaviour (#7507)
* docs: update allowed mentions object section

Updated the section with the new components v2 behaviour.

* docs: update create message and edit message section

Updated both sections with the new components v2 behaviour.

* docs: update sending a message with a component section

Added a warning alert box to make users aware of the "IS_COMPONENTS_V2" message flag being irreversible when set to a message.

* docs: update execute webhook and edit webhook message section

Updated both desction with the new components v2 behaviour.

* docs: update edit message section

Updated the section with the asterisk being moved from the "flags" description to the "flags" field. Removed one of the asterisks since only one was present inside the table.

* fix: reverted change to avoid conflicts with #7508

* fixed table