Revert "add `unusual_dm_activity_until` to member object (#7650)"
This reverts commit 1520a812cdfebc74098e1e73aa79344be16af9b0.

Apolgies folks - I looked up the wrong attribute on the PR.
Rolling back while we work out if this should be enabled as
public or not. Totally my fault for not looking properly.

The version of the docs platform will make these issues far
less of a problem (or 🤞🏻 get rid of them entirely).